+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/bot/Ktor/KtorRequestsExecutor.kt
@@ -5,6 +5,7 @@ import dev.inmo.tgbotapi.bot.BaseRequestsExecutor
 import dev.inmo.tgbotapi.bot.Ktor.base.*
 import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.bot.exceptions.newRequestException
+import dev.inmo.tgbotapi.bot.ktor.KtorPipelineStepsHolder
 import dev.inmo.tgbotapi.bot.settings.limiters.ExceptionsOnlyLimiter
 import dev.inmo.tgbotapi.bot.settings.limiters.RequestLimiter
 import dev.inmo.tgbotapi.requests.abstracts.Request
@@ -56,7 +57,8 @@ class KtorRequestsExecutor(
     callsFactories: List<KtorCallFactory> = emptyList(),
     excludeDefaultFactories: Boolean = false,
     private val requestsLimiter: RequestLimiter = ExceptionsOnlyLimiter(),
-    private val jsonFormatter: Json = nonstrictJsonFormat
+    private val jsonFormatter: Json = nonstrictJsonFormat,
+    private val pipelineStepsHolder: KtorPipelineStepsHolder = KtorPipelineStepsHolder
 ) : BaseRequestsExecutor(telegramAPIUrlsKeeper) {
     private val callsFactories: List<KtorCallFactory> = callsFactories.run {
         if (!excludeDefaultFactories) {
@@ -73,37 +75,51 @@ class KtorRequestsExecutor(
     }
 
     override suspend fun <T : Any> execute(request: Request<T>): T {
-        return safely(
-            { e ->
-                throw if (e is ClientRequestException) {
-                    val content = e.response.readText()
-                    val responseObject = jsonFormatter.decodeFromString(Response.serializer(), content)
-                    newRequestException(
-                        responseObject,
-                        content,
-                        "Can't get result object from $content"
-                    )
-                } else {
-                    e
-                }
-            }
-        ) {
-            requestsLimiter.limit {
-                var result: T? = null
-                for (potentialFactory in callsFactories) {
-                    result = potentialFactory.makeCall(
-                        client,
-                        telegramAPIUrlsKeeper,
-                        request,
-                        jsonFormatter
-                    )
-                    if (result != null) {
-                        break
-                    }
-                }
+        return runCatching {
+            safely(
+                { e ->
+                    pipelineStepsHolder.onRequestException(request, e) ?.let { return@safely it }
 
-                result ?: error("Can't execute request: $request")
+                    throw if (e is ClientRequestException) {
+                        val content = e.response.readText()
+                        val responseObject = jsonFormatter.decodeFromString(Response.serializer(), content)
+                        newRequestException(
+                            responseObject,
+                            content,
+                            "Can't get result object from $content"
+                        )
+                    } else {
+                        e
+                    }
+
+                }
+            ) {
+                pipelineStepsHolder.onBeforeSearchCallFactory(request, callsFactories)
+                requestsLimiter.limit {
+                    var result: T? = null
+                    lateinit var factoryHandledRequest: KtorCallFactory
+                    for (potentialFactory in callsFactories) {
+                        pipelineStepsHolder.onBeforeCallFactoryMakeCall(request, potentialFactory)
+                        result = potentialFactory.makeCall(
+                            client,
+                            telegramAPIUrlsKeeper,
+                            request,
+                            jsonFormatter
+                        )
+                        result = pipelineStepsHolder.onAfterCallFactoryMakeCall(result, request, potentialFactory)
+                        if (result != null) {
+                            factoryHandledRequest = potentialFactory
+                            break
+                        }
+                    }
+
+                    result ?.let {
+                        pipelineStepsHolder.onRequestResultPresented(it, request, factoryHandledRequest, callsFactories)
+                    } ?: pipelineStepsHolder.onRequestResultAbsent(request, callsFactories) ?: error("Can't execute request: $request")
+                }
             }
+        }.let {
+            pipelineStepsHolder.onRequestReturnResult(it, request, callsFactories)
         }
     }