+++ tgbotapi.core/src/jvmMain/kotlin/dev/inmo/tgbotapi/utils/StorageFileFactory.kt
@@ -7,7 +7,7 @@ import java.nio.file.Files
 fun StorageFile(
     file: File
 ) = StorageFile(
-    StorageFileInfo(file.name)
+    file.name
 ) {
     file.inputStream().asInput()
 }